Adding authentication form types
