+++ Authentication/SignUp/View/SignupView.swift
@@ -6,56 +6,7 @@
 //  Copyright © 2016 Wolox. All rights reserved.
 //
 
-import Foundation
-
-public protocol SignupViewType: Renderable {
- 
-    var view: UIView { get }
-    
-    var usernameLabel: UILabel? { get }
-    var usernameTextField: UITextField? { get }
-    var usernameValidationMessageLabel: UILabel? { get }
-    
-    var emailLabel: UILabel? { get }
-    var emailTextField: UITextField { get }
-    var emailValidationMessageLabel: UILabel? { get }
-    
-    var passwordLabel: UILabel? { get }
-    var passwordTextField: UITextField { get }
-    var passwordValidationMessageLabel: UILabel? { get }
-    var passwordVisibilityButton: UIButton? { get }
-    
-    var passwordConfirmLabel: UILabel? { get }
-    var passwordConfirmTextField: UITextField { get }
-    var passwordConfirmValidationMessageLabel: UILabel? { get }
-    var passwordConfirmVisibilityButton: UIButton? { get }
-    
-    var signupButton: UIButton { get }
-    var signupErrorLabel: UILabel? { get }
-    
-    var termsAndServicesButton: UIButton { get }
-    var termsAndServicesLabel: UILabel? { get }
-    
-    var usernameTextFieldValid: Bool { get set }
-    var usernameTextFieldSelected: Bool { get set }
-    var emailTextFieldValid: Bool { get set }
-    var emailTextFieldSelected: Bool { get set }
-    var passwordTextFieldValid: Bool { get set }
-    var passwordTextFieldSelected: Bool { get set }
-    var passwordConfirmationTextFieldValid: Bool { get set }
-    var passwordConfirmationTextFieldSelected: Bool { get set }
-    var signupButtonEnabled: Bool { get set }
-    var signupButtonPressed: Bool { get set }
-    
-}
-
-public extension SignupViewType where Self: UIView {
-    
-    var view: UIView {
-        return self
-    }
-    
-}
+public protocol SignupViewType: Renderable, SignupFormType { }
 
 public final class SignupView: UIView, SignupViewType {
     
@@ -92,8 +43,10 @@ public final class SignupView: UIView, SignupViewType {
     public var emailTextFieldSelected: Bool = false
     public var passwordTextFieldValid: Bool = false
     public var passwordTextFieldSelected: Bool = false
+    public var showPassword: Bool = false
     public var passwordConfirmationTextFieldValid: Bool = false
     public var passwordConfirmationTextFieldSelected: Bool = false
+    public var showConfirmPassword: Bool = false
     public var signupButtonEnabled: Bool = false
     public var signupButtonPressed: Bool = false
